chore: update main branch

+++ src/main/java/dev/gigafyde/apollo/core/TrackScheduler.java
@@ -9,12 +9,6 @@ import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrackEndReason;
 import dev.gigafyde.apollo.utils.SongUtils;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
-import java.util.Queue;
-import java.util.concurrent.LinkedBlockingDeque;
 import lavalink.client.player.IPlayer;
 import lavalink.client.player.LavalinkPlayer;
 import lavalink.client.player.event.PlayerEventListenerAdapter;
@@ -23,11 +17,16 @@ import net.dv8tion.jda.api.entities.TextChannel;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.*;
+import java.util.concurrent.LinkedBlockingDeque;
+
 public class TrackScheduler extends PlayerEventListenerAdapter {
     private static final Logger log = LoggerFactory.getLogger("TrackScheduler");
     private final LavalinkPlayer player;
     private final AudioPlayerManager audioPlayerManager;
     private boolean looped;
+    private boolean announceTrack = true;
+    private boolean announceLoop = true;
     private TextChannel boundChannel;
     private Message nowPlaying;
     private AudioTrack loopedTrack;
@@ -66,16 +65,25 @@ public class TrackScheduler extends PlayerEventListenerAdapter {
             return;
 
         player.playTrack(track);
-        if (player.getPlayingTrack() == track) setPreviousTrack(track);
 
         if (boundChannel != null) {
             try {
+                // if announcements are disabled stop code execution
+                if (!announceTrack || !announceLoop && looped) return;
                 // Try to delete the previous now-playing message
                 nowPlaying.delete().complete();
             } catch (Exception ignored) {
                 // If it fails. it'll most likely be because of something on discord's end. so it's not our problem.
             }
             if (player.getPlayingTrack() == track)
+
+				/*
+				If this is null, it should mean we only just created a new scheduler
+		        Sending a now-playing image in this case is overkill as we already sent a added to queue image
+				 */
+	            if (previousTrack == null) return;
+
+
                 try {
                     boundChannel.sendFile(Objects.requireNonNull(SongUtils.generateNowPlaying(track, 1)), "nowplaying.png").queue(msg -> nowPlaying = msg);
                 } catch (Exception e) {
@@ -120,6 +128,22 @@ public class TrackScheduler extends PlayerEventListenerAdapter {
 
     public void setLooped(boolean active) {
         looped = active;
+    }
+
+    public boolean isAnnounceTrack() {
+        return announceTrack;
+    }
+
+    public void setAnnounceTrack(boolean active) {
+        announceTrack = active;
+    }
+
+    public boolean isAnnounceLoop() {
+        return announceLoop;
+    }
+
+    public void setAnnounceLoop(boolean active) {
+        announceLoop = active;
     }
 
 //    public int addTracks(String author, AudioTrack... tracks) {
